Attempt to create a JointAcccountFactory
This commit attempts to create a JointAccountFactory. The intent was
to create a JointAccount contract per shared account. This type of
functionality works in other VMs, but it does NOT work in Vite VM.

Charles Liu — 03/08/2022 on Vite discord #general-dev:

Unfortunately it's impossible to create a contract from another contract
using new keyword in any release version of Solidity++ so far. We're
working on Vite VM to enable the create instruction.

https://discordapp.com/channels/425843004616474635/425844259745300481/950802614952538192